learner-pwa/ml-service/requirements.txt
# Core ML/AI Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.4.0
tensorflow==2.15.0
numpy==1.26.3
pandas==2.1.4

# Model Management
mlflow==2.9.2
joblib==1.3.2

# Data Processing
scipy==1.11.4

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
